==> docs/02-ProjectOrganization.Rmd <==
# Project Organization {#projorg}

Each publication or dataset you work with will need its own R project to contain all of the work you will do to digitize the model and get it ready to be added to the PADRINO database. More information about selecting a publication to digitize and initializing a project can be found in Chapter \@ref(genwf) : General Workflow.

All of your projects should be kept together in a folder. Therefore, it is important that you title your R projects consistently and descriptively. Generally, using the first author's last name and the year of publication separated by an underscore is most appropriate. There may be some instances where appending more information, like the journal or the focal species, is necessary. 

Your folder of projects should be synced [SHARED CLOUD SERVICE]. This gives us all access to your scripts as you are working on them, which will be helpful when you are working remotely.


## Folder organization

Within the project home folder, you should create three subfolders:

  * 'R', which will contain all of your scripts  
  * 'data', which will contain all data (more on this in a moment)  
  * 'results', which will contain any figures you create  

You may have other files that you want to keep in a project, such as a PDF of the full article text, supplemental files/appendices, etc. For the most part, these can also be kept in the project home folder, EXCEPT if they contain readable data files (Excel tables or .csv, etc). Any data files that you might use in your digitization should go into the 'data' folder, otherwise there is no reason to include them in the R project. Depending on the publication you're working with, the amount of extra files you want to keep on-hand might be quite a lot. In that case, you should create a fourth subfolder titled 'supplement' and store everything in there.

### R folder
Scripts in the 'R' folder should be given names beginning with numbers indicating the order in which they should be executed. The first script will almost always be something like "01_DataFormatting.R" in which you will prepare the data you downloaded or otherwise extracted to get it ready for modeling. The rest of the scripts you might write are detailed in Chapter \@ref(genwf) : General Workflow. The file names suggested in that script can be modified by adding additional details about the dataset you are working with, if it helps you to better keep track of your work. Whatever you decide, please try to be consistent within each project! 

### Data folder
The data folder contains both the files with the original data, and the output of formatting that same data. Always be sure to use descriptive names so you don't mix up the raw data and the new data you worked so hard to put together!

### Results folder
Use the results folder to store the plots you need to make to check data before or during model fitting. To save yourself the hassle of recreating plots every time someone needs to take a look at them, export them each time you update them.

### Example of folder organization
<center>
```{r echo=FALSE, message=FALSE, warning=FALSE}
library( htmlTable )
library( tidyverse )

folder_ex <- matrix( data = "", nrow = 27, ncol = 3 )
folder_ex[1,1] <- "&#128193; Author_year_species"
folder_ex[2,2] <- '<span style="color:dodgerblue;">&#127297;</span> Author_year_species.Rproj'
folder_ex[3,2] <- "&#128193; data"
folder_ex[4,3] <- '<span style="color:forestgreen;">&#127303;</span> raw_data.csv'
folder_ex[5,3] <- '<span style="color:forestgreen;">&#127303;</span> survival_data.csv'
folder_ex[6,3] <- '<span style="color:forestgreen;">&#127303;</span> growth_data.csv'
folder_ex[7,3] <- '<span style="color:forestgreen;">&#127303;</span> recruitment_data.csv'
folder_ex[8,2] <- "&#128193; R"
folder_ex[9,3] <- '<span style="color:dodgerblue;">&#9415;</span> 01_DataFormatting.R'
folder_ex[10,3] <- '<span style="color:dodgerblue;">&#9415;</span> 02_Plotting.R'
folder_ex[11,3] <- '<span style="color:dodgerblue;">&#9415;</span> 03_VitalRates.R'
folder_ex[12,3] <- '<span style="color:dodgerblue;">&#9415;</span> 04_IPMScratch.R'
folder_ex[13,3] <- '<span style="color:dodgerblue;">&#9415;</span> 05_ipmr.R'
folder_ex[14,3] <- '<span style="color:dodgerblue;">&#9415;</span> 06_PADRINO.R'
folder_ex[15,2] <- "&#128193; results"
folder_ex[16,3] <- "&#128443; survival_plot.png"
folder_ex[17,3] <- "&#128443; growth_plot.png"
folder_ex[18,3] <- "&#128443; recruitment_plot.png"
folder_ex[19,2] <- "&#128193; supplement"
folder_ex[20,3] <- "&#128442; Author_year_Appendix1.pdf"
folder_ex[21,3] <- "&#128442; Author_year_Appendix2.pdf"
folder_ex[22,2] <- "&#128193; temp"
folder_ex[23,3] <- '<span style="color:forestgreen;">&#127303;</span> test_data.csv'
folder_ex[24,2] <- "&#128442; Author_year_fulltext.pdf"
folder_ex[25,2] <- '<span style="color:forestgreen;">&#127303;</span> Author_year_species_pdb.xlsx'
folder_ex[26,2] <- '<span style="color:dodgerblue;">&#9415;</span> Author_year_species_wrapper.R'
folder_ex[27,2] <- "&#128441; Author_year_species_notes.txt"

folder_ex[c(4,9,16,20,23),2] <- "     &#8627;"
folder_ex[2,1] <- "     &#8627;"

folder_ex %>% htmlTable( align = "lll" )
```
</center>



## More complicated datasets

For some datasets, the guidelines provided here are fully comprehensive. But some datasets may not be so simple, and they might require you to work on only a subset of the data at a time.

In general, if it is a case where you are splitting the dataset into two subsets, or you are working with two slightly different models, it is fine to just add an appendix to your file names. Examples of when this is the case include:

  * A model excluding the seedbank, and a model including the seedbank:  
    + 03_IPMScratch_noseedbank.R  
    + 03_IPMScratch_seedbank.R  
      
  * Two species (abbreviated with first 3 letters of genus & species):  
    + 03_IPMScratch_Andger.R  
    + 03_IPMScratch_Boucur.R  

However, some datasets may necessitate the use of more than two different models or subsets. In these cases, it is best to create separate subfolders. Whether these subfolders go into the project home folder, or into the standard subfolders, it depends on the raw data. If you download more than one raw data file corresponding to each IPM (or set of IPMs), then the new subfolders should go into the project home folder. Ex.:
<center>
```{r echo=FALSE, message=FALSE, warning=FALSE}
folder_ex2 <- matrix( data = "", nrow = 16, ncol = 4)
folder_ex2[1,1] <- "&#128193; Author_year"
folder_ex2[2,1] <- "     &#8627;"
folder_ex2[2,2] <- "&#128193; SpeciesA"
folder_ex2[3,3] <- "&#128193; data"
folder_ex2[4,4] <- '<span style="color:forestgreen;">&#127303;</span> raw_data_speciesA.csv'
folder_ex2[5,3] <- "&#128193; R"
folder_ex2[6,3] <- "&#128193; results"
folder_ex2[7,2] <- "&#128193; SpeciesB"
folder_ex2[8,3] <- "&#128193; data"
folder_ex2[9,4] <- '<span style="color:forestgreen;">&#127303;</span> raw_data_speciesB.csv'
folder_ex2[10,3] <- "&#128193; R"
folder_ex2[11,3] <- "&#128193; results"
folder_ex2[12,2] <- "&#128193; SpeciesB"
folder_ex2[13,3] <- "&#128193; data"
folder_ex2[14,4] <- '<span style="color:forestgreen;">&#127303;</span> raw_data_speciesB.csv'
folder_ex2[15,3] <- "&#128193; R"
folder_ex2[16,3] <- "&#128193; results"

folder_ex2[c(3,8,13),2] <- "     &#8627;"
folder_ex2[c(4,9,14),3] <- "     &#8627;"

folder_ex2 %>% htmlTable(align = "lll")
```
</center>



Conversely, if the raw data are all contained within a single file (or multiple files all containing some kind of data for every model), then the subfolders should be nested within the standard subfolders. This is because you may have some scripts which apply to all versions of the model, like the first data formatting script, but then you will write different scripts for each version of the model that should be kept separate. Ex.:
<center>
```{r echo=FALSE, message=FALSE, warning=FALSE}
folder_ex3 <- matrix( data = "", nrow = 15, ncol = 3)
folder_ex3[1,1] <- "&#128193; Author_year"
folder_ex3[2,1] <- "     &#8627;"
folder_ex3[2,2] <- "&#128193; data"
folder_ex3[3,3] <- '<span style="color:forestgreen;">&#127303;</span> raw_data_all_species.csv'
folder_ex3[4,3] <- "&#128193; SpeciesA"
folder_ex3[5,3] <- "&#128193; SpeciesB"
folder_ex3[6,3] <- "&#128193; SpeciesC"
folder_ex3[7,2] <- "&#128193; R"
folder_ex3[8,3] <- '<span style="color:dodgerblue;">&#9415;</span> 01_DataFormatting.R'
folder_ex3[9,3] <- "&#128193; SpeciesA"
folder_ex3[10,3] <- "&#128193; SpeciesB"
folder_ex3[11,3] <- "&#128193; SpeciesC"
folder_ex3[12,2] <- "&#128193; results"
folder_ex3[13,3] <- "&#128193; SpeciesA"
folder_ex3[14,3] <- "&#128193; SpeciesB"
folder_ex3[15,3] <- "&#128193; SpeciesC"

folder_ex3[c(3,8,13),2] <- "     &#8627;"

folder_ex3 %>% htmlTable(align = "lll")
```
</center>



Ultimately, the decision to split into different subfolders and how to nest these folders is hard to make *a priori*. You might discover along the way that you need to simplify the model before tackling the whole thing, and you might end up making way more than two versions of a model. It is worthwhile to spend some extra time reorganizing your project if you think it will make it easier to understand.

Finally, always be sure to explain any project structure/nestedness in the wrapper file (more details in Chapter \@ref(genwf) : General Workflow).
